requirements.txt
celery
SQLAlchemy
tornado
requests
hashlib
Pillow